+++ objectbox/lib/src/native/sync.dart
@@ -175,8 +175,7 @@ class SyncChange {
   SyncChange._(this.entityId, this.entity, this.puts, this.removals);
 }
 
-/// Sync client is used to connect to an ObjectBox sync server.
-/// Use through [Sync].
+/// A Sync client is used to connect to an ObjectBox Sync server.
 class SyncClient {
   final Store _store;
 
@@ -189,9 +188,47 @@ class SyncClient {
       : throw StateError('SyncClient already closed');
 
   /// Creates a Sync client associated with the given store and options.
-  /// This does not initiate any connection attempts yet: call start() to do so.
-  SyncClient._(this._store, List<String> serverUrls,
-      List<SyncCredentials> credentials, Map<String, String>? filterVariables) {
+  /// This does not initiate any connection attempts yet, call [start] to do so.
+  ///
+  /// By default, a Sync client automatically receives updates from the server
+  /// once login succeeded. To configure this differently, call
+  /// [setRequestUpdatesMode] with the wanted mode.
+  ///
+  /// Typically, only a single URL and a single credentials object is passed:
+  ///
+  /// ```dart
+  /// // Connect to a test server running on localhost using
+  /// // an unencrypted connection without authentication
+  /// SyncClient client =
+  ///     SyncClient(store, ['ws://127.0.0.1:9999'], [SyncCredentials.none()]);
+  /// ```
+  ///
+  /// Passing multiple URLs allows high availability and load balancing (for ex.
+  /// using a ObjectBox Sync Server Cluster). A random URL is selected for each
+  /// connection attempt.
+  ///
+  /// When passing multiple credentials, can't include [SyncCredentials.none].
+  ///
+  /// To configure [Sync filter](https://sync.objectbox.io/sync-server/sync-filters)
+  /// variables, pass variable names mapped to their value to [filterVariables].
+  ///
+  /// Sync client filter variables can be used in server-side Sync filters to
+  /// filter out objects that do not match the filter.
+  ///
+  /// To, for example, use self-signed certificates in a local development
+  /// environment or custom CAs, pass certificate paths referring to the local
+  /// file system to [certificatePaths].
+  ///
+  /// To configure Sync behavior, pass bitwise OR-ed [OBXSyncFlags] values to
+  /// [flags]. See [OBXSyncFlags] for available flags.
+  SyncClient(
+      this._store, List<String> serverUrls, List<SyncCredentials> credentials,
+      {Map<String, String>? filterVariables,
+      List<String>? certificatePaths,
+      int? flags}) {
+    if (syncClientsStorage.containsKey(_store)) {
+      throw StateError('Only one sync client can be active for a store');
+    }
     if (serverUrls.isEmpty) {
       throw ArgumentError.value(
           serverUrls, "serverUrls", "Provide at least one server URL");
@@ -203,11 +240,32 @@ class SyncClient {
           'Please visit https://objectbox.io/sync/ for options.');
     }
 
-    _cSync = withNativeStrings(
-        serverUrls,
-        (ptr, size) => checkObxPtr(
-            C.sync_urls(InternalStoreAccess.ptr(_store), ptr, size),
-            'failed to create Sync client'));
+    // Build options
+    final options = checkObxPtr(C.sync_opt(InternalStoreAccess.ptr(_store)),
+        'failed to create Sync options');
+
+    for (final url in serverUrls) {
+      withNativeString(url, (urlCStr) {
+        checkObx(C.sync_opt_add_url(options, urlCStr));
+      });
+    }
+
+    if (certificatePaths != null) {
+      for (final certPath in certificatePaths) {
+        withNativeString(certPath, (certPathCStr) {
+          checkObx(C.sync_opt_add_cert_path(options, certPathCStr));
+        });
+      }
+    }
+
+    // Note: 0 or invalid flags are ignored by sync_opt_flags
+    if (flags != null) {
+      checkObx(C.sync_opt_flags(options, flags));
+    }
+
+    // Create Sync client with options (options are freed by sync_create)
+    _cSync =
+        checkObxPtr(C.sync_create(options), 'failed to create Sync client');
 
     filterVariables?.forEach(putFilterVariable);
 
@@ -217,6 +275,9 @@ class SyncClient {
       // also covers the length == 0 case
       setMultipleCredentials(credentials);
     }
+
+    syncClientsStorage[_store] = this;
+    InternalStoreAccess.addCloseListener(_store, this, close);
   }
 
   /// Closes and cleans up all resources used by this sync client.
@@ -719,7 +780,7 @@ class _SyncListenerGroup<StreamValueType> {
 /// [ObjectBox Sync](https://objectbox.io/sync/) makes data available and
 /// synchronized across devices, online and offline.
 ///
-/// Start a client using [Sync.client()] and connect to a remote server.
+/// Create a client using the [SyncClient] constructor.
 class Sync {
   /// Set to `true` to enable shared global IDs for a Sync-enabled entity class.
   ///
@@ -745,56 +806,81 @@ class Sync {
   /// Returns true if the loaded ObjectBox native library supports Sync.
   static bool isAvailable() => _syncAvailable;
 
-  /// Creates a Sync client associated with the given store and configures it
+  /// Creates a [SyncClient] associated with the given store and configures it
   /// with the given options. This does not initiate any connection attempts
-  /// yet, call [SyncClient.start()] to do so.
+  /// yet, call [SyncClient.start] to do so.
   ///
-  /// Before [SyncClient.start()], you can still configure some aspects of the
-  /// client, e.g. its [SyncRequestUpdatesMode] mode.
+  /// By default, a Sync client automatically receives updates from the server
+  /// once login succeeded. To configure this differently, call
+  /// [SyncClient.setRequestUpdatesMode] with the wanted mode.
   ///
   /// To configure [Sync filter](https://sync.objectbox.io/sync-server/sync-filters)
   /// variables, pass variable names mapped to their value to [filterVariables].
   ///
   /// Sync client filter variables can be used in server-side Sync filters to
   /// filter out objects that do not match the filter.
+  ///
+  /// To, for example, use self-signed certificates in a local development
+  /// environment or custom CAs, pass certificate paths referring to the local
+  /// file system to [certificatePaths].
+  ///
+  /// To configure Sync behavior, pass bitwise OR-ed [OBXSyncFlags] values to
+  /// [flags]. See [OBXSyncFlags] for available flags.
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient client(
           Store store, String serverUrl, SyncCredentials credentials,
-          {Map<String, String>? filterVariables}) =>
-      clientMultiUrls(store, [serverUrl], credentials,
-          filterVariables: filterVariables);
+          {Map<String, String>? filterVariables,
+          List<String>? certificatePaths,
+          int? flags}) =>
+      SyncClient(store, [serverUrl], [credentials],
+          filterVariables: filterVariables,
+          certificatePaths: certificatePaths,
+          flags: flags);
 
   /// Like [client], but accepts a list of credentials.
   ///
   /// When passing multiple credentials, does **not** support
-  /// [SyncCredentials.none()].
+  /// [SyncCredentials.none].
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient clientMultiCredentials(
           Store store, String serverUrl, List<SyncCredentials> credentials,
-          {Map<String, String>? filterVariables}) =>
-      clientMultiCredentialsMultiUrls(store, [serverUrl], credentials,
-          filterVariables: filterVariables);
+          {Map<String, String>? filterVariables,
+          List<String>? certificatePaths,
+          int? flags}) =>
+      SyncClient(store, [serverUrl], credentials,
+          filterVariables: filterVariables,
+          certificatePaths: certificatePaths,
+          flags: flags);
 
   /// Like [client], but accepts a list of URLs to work with multiple servers.
+  ///
+  /// Passing multiple URLs allows high availability and load balancing (for ex.
+  /// using a ObjectBox Sync Server Cluster). A random URL is selected for each
+  /// connection attempt.
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient clientMultiUrls(
           Store store, List<String> serverUrls, SyncCredentials credentials,
-          {Map<String, String>? filterVariables}) =>
-      clientMultiCredentialsMultiUrls(store, serverUrls, [credentials],
-          filterVariables: filterVariables);
+          {Map<String, String>? filterVariables,
+          List<String>? certificatePaths,
+          int? flags}) =>
+      SyncClient(store, serverUrls, [credentials],
+          filterVariables: filterVariables,
+          certificatePaths: certificatePaths,
+          flags: flags);
 
   /// Like [client], but accepts a list of credentials and a list of URLs to
   /// work with multiple servers.
   ///
   /// When passing multiple credentials, does **not** support
-  /// [SyncCredentials.none()].
-  static SyncClient clientMultiCredentialsMultiUrls(
-      Store store, List<String> serverUrls, List<SyncCredentials> credentials,
-      {Map<String, String>? filterVariables}) {
-    if (syncClientsStorage.containsKey(store)) {
-      throw StateError('Only one sync client can be active for a store');
-    }
-    final client =
-        SyncClient._(store, serverUrls, credentials, filterVariables);
-    syncClientsStorage[store] = client;
-    InternalStoreAccess.addCloseListener(store, client, client.close);
-    return client;
-  }
+  /// [SyncCredentials.none].
+  @Deprecated('Use the SyncClient constructor instead')
+  static SyncClient clientMultiCredentialsMultiUrls(Store store,
+          List<String> serverUrls, List<SyncCredentials> credentials,
+          {Map<String, String>? filterVariables,
+          List<String>? certificatePaths,
+          int? flags}) =>
+      SyncClient(store, serverUrls, credentials,
+          filterVariables: filterVariables,
+          certificatePaths: certificatePaths,
+          flags: flags);
 }